Exchange build-dist.sh shell script for a pyproject-based task executor.
I've found `poethepoet` to be great replacement for Makefiles or shell
scripts in python projects. The first advantage is the rules are
included in the pyproject.toml file instead of an ancillary file
(fewer things to keep track of). The second is poe is very versatile,
being able to run python code directly in addition to commands like a
shell script. With this commit, I've added poethepoet as a development
dependency and and replaced the build-dist.sh script with a poe task
called "bundle" that replicates the functionality of the script
exactly.